cleanup, split request further into request_response.go and request_sendrawtx.go
because it's getting way too big
